API Documentation

Lists of docstrings for those that are documented.

API Documentation

StructureFrames and related constructs

Hydrogen bonding model

Frame geometry

Internal geometry constructs

Related constants

StructureFrames and related constructs

ProtStructs.StructureFrame — Type

StructureFrame

Overall model for protein structure. Contains information on the atomic coordinates, atom information, and the residues. Is obtained from using extractframe on a Chemfiles Frame object.

ProtStructs.extractframe — Function

extractframe(fr::Chemfiles.Frame)

Extract Chemfiles.Frame to get objects of structs defined in the package. Note that not all properties will be inherited/converted.

ProtStructs.JResidue — Type

JResidue

A residue in a probed frame. Contains the residue name. Holds a list of indices of atoms in the StructureFrame. Other properties include the chainid specifying the chain in which the residue belongs and whether the residue is included in the described protein structure.

ProtStructs.JAtom — Type

JAtom{T<:AbstractFloat}

Represents an atom in a probed frame. Contains basic information such as its name and type as given in the PDB (input file). Mass and charge are obtained from Chemfiles inference given atom type.

Note: charge of an atom is e * jatom.charge with e as elementary charge.

ProtStructs.getatom — Function

getatom(residue, name::Symbol) -> UInt64

Get the index of an atom in the parent frame of the residue that has the name provided.

ProtStructs.setatom! — Function

setatom!(residue, atom::JAtom, idx)

Sets an atom to a residue.

Note: This does not check if the atom with a given .name is already in the residue and instead mutates it.

ProtStructs.addprotons! — Function

addprotons!(fr::StructureFrame[; has_na])

Add backbone protons to residues in the Frame fr. Estimate the position of the missing proton using the method used by DSSP (within its source code). Optional keyword argument has_na indicates whether the frame fr has nucleic acids added in parsing. Defaults to false.

Note: Setting has_na is important, when it can be shown that the frame has nucleic acids.

ProtStructs.getchaindict — Function

getchaindict(frame::StructureFrame)

Get a dictionary of chains in the frame containing UnitRanges of standard protein residues.

Hydrogen bonding model

ProtStructs.HBondDict — Type

HBondDict{N,T<:AbstractFloat}

Hydrogen bonding dictionary for a specific residue. Type parameter N describes at most how many hydrogen bonds can be donated/accepted by the amide part of the residue. Index of the residue (in a given topology) is this.handle.

Note: Values of N that make sense are $1$ and $2$; any greater would not quite make sense. DSSP uses two to store the H-bonds with lowest energy (most stable).

ProtStructs.recorddonor! — Function

recorddonor!(hbdict, i, bondenergy)
recordacceptor!(hbdict, i, bondenergy)

Record a donor/acceptor to the hydrogen bonding dictionary. The bond is with the residue at index i. The corresponding bond energy is bondenergy.

Using these functions records entries with increasing energy (decreasing bond strength).

ProtStructs.hbondenergy — Function

hbondenergy(frame, donor, acceptor; method=:dist)

Calculate the (amide) hydrogen bond energy between a donor residue and an acceptor residue in a frame. Both nonframe arguments are indices in the StructureFrame. Defaults to a mimicked method of that currently used in DSSP (using distances, although there are plans to use angles). An approximate potential using dot products method=:dot is provided in this module.

Keyword arguments

method: Defaults to :dist. Method used for energy calculation.

ProtStructs.hbonded — Function

hbonded(hbdicts, res_i, res_j) -> Bool

Check if the residue at index res_i acts as a hydrogen bond acceptor to the residue at res_j based on the list of dictionaries hbdicts. Equivalently, if residue at res_j acts as a hydrogen bond donor to the residue at res_i.

ProtStructs.nturn — Function

nturn(hbdicts, i, n) -> Bool
alphaturn(hbdicts, i) = nturn(hbdicts, i, 4)
shortturn(hbdicts, i) = nturn(hbdicts, i, 3)
piturn(hbdicts, i) = nturn(hbdicts, i, 5)

Check if there is an $n$-turn starting from residue with index i. Uses the vector hbdicts as reference. DSSP defines turns with specific n such as $3,4,5$-turns. These are shortturn, alphaturn, and piturn, respectively, in this module.

ProtStructs.parallelbridge — Function

parallelbridge(hbdicts, i, j)
antiparallelbridge(hbdicts, i, j)

Check if there is an elementary bridge between the residue with indices i and j. Check the original DSSP reference (link in README) or the Julia source codes for the definitions of these bridges.

Frame geometry

ProtStructs.phi — Function

phi(frame, i[, chaindict]) ∈ [-pi, pi], Nothing

Get the phi ($ϕ$) angle of the residue in the frame pointed by index i. Returns nothing if cannot be defined.

Use the chain dictionary chaindict as reference. This can be supplied if previously calculated, so there'd be no need to calculate for each index. Defaults to the builtin getchaindict function.

Note: It is recommended to use only when all indices i can be found in the chain dictionary. Furthermore, it is also recommended to make sure chains are those of amino acid residues.

ProtStructs.psi — Function

psi(frame, i[, chaindict]) ∈ [-pi, pi], Nothing

Get the psi ($ψ$) angle of the residue in the frame pointed by index i. Returns nothing if cannot be defined.

Use the chain dictionary chaindict as reference. This can be supplied if previously calculated, so there'd be no need to calculate for each index. Defaults to the builtin getchaindict function.

Note: It is recommended to use only when all indices i can be found in the chain dictionary. Furthermore, it is also recommended to make sure chains are those of amino acid residues.

ProtStructs.ramachandranangles — Function

ramachandranangles(frame, i[, chaindict]) -> Tuple
ProtStructs.ϕψ(frame, i[, chaindict])

Get the Ramachandran angles of the residue in the frame pointed by index i. Uses phi and psi functions as defined. Returns a tuple based on that order. Shorthand above is provided, but not exported.

Returns nothing if cannot be defined.

Use the chain dictionary chaindict as reference. This can be supplied if previously calculated, so there'd be no need to calculate for each index. Defaults to the builtin getchaindict function.

Note: It is recommended to use only when all indices i can be found in the chain dictionary. Furthermore, it is also recommended to make sure chains are those of amino acid residues.

ProtStructs.omega — Function

omega(frame, i[, chaindict]) ∈ [-pi, pi], Nothing

Get the omega ($ω$) angle of the residue in the frame pointed by index i. Returns nothing if cannot be defined.

Use the chain dictionary chaindict as reference. This can be supplied if previously calculated, so there'd be no need to calculate for each index. Defaults to the builtin getchaindict function.

Note: It is recommended to use only when all indices i can be found in the chain dictionary. Furthermore, it is also recommended to make sure chains are those of amino acid residues.

ProtStructs.kappa — Function

kappa(frame, i[, chaindict]) ∈ [0, pi], Nothing

Get the kappa ($κ$) angle of the residue in the fram pointed by index i. Returns nothing if cannot be defined.

Use the chain dictionary chaindict as reference. This can be supplied if previously calculated, so there'd be no need to calculate for each index. Defaults to the builtin getchaindict function.

Note: It is recommended to use only when all indices i can be found in the chain dictionary. Furthermore, it is also recommended to make sure chains are those of amino acid residues.

ProtStructs.alpha — Function

alpha(frame, i[, chaindict]) ∈ [-pi, pi], Nothing
chirality(frame, i, chaindict) ∈ (-1, 0, 1), Nothing

Get the alpha ($α$) angle of the residue in the frame pointed by index i. Chirality (in HSSP) is defined as the sign of alpha. Returns nothing if cannot be defined.

Use the chain dictionary chaindict as reference. This can be supplied if previously calculated, so there'd be no need to calculate for each index. Defaults to the builtin getchaindict function.

Note: It is recommended to use only when all indices i can be found in the chain dictionary. Furthermore, it is also recommended to make sure chains are those of amino acid residues.

ProtStructs.hsspangles — Function

hsspangles(frame, i[, chaindict]) -> Tuple
ProtStructs.κα(frame, i[, chaindict])

Get kappa and alpha angles for the residue in the frame pointed by index i. These angles are used by HSSP in searching for homologous structures. Shorthand above is provided, but not exported.

Returns nothing if cannot be defined.

Use the chain dictionary chaindict as reference. This can be supplied if previously calculated, so there'd be no need to calculate for each index. Defaults to the builtin getchaindict function.

Note: It is recommended to use only when all indices i can be found in the chain dictionary. Furthermore, it is also recommended to make sure chains are those of amino acid residues.

Internal geometry constructs

ProtStructs.distance — Function

distance(a, b) -> Real

Calculate Euclidean distance from a to b.

ProtStructs.anglespan — Function

anglespan(a, b, c) ∈ [0, π]

Calculate smaller angle between two vectors $AB$ and $BC$ meeting at $B$.

ProtStructs.dihedral — Function

dihedral(a, b, c, d) ∈ [-π, π]

Calculate right handed dihedral (torsion) angle between the plane $ABC$ and $BCD$ intersecting along $BC$. $ABC$ is determined by the first three arguments as points, while $BCD$ is determined by the last three.

ProtStructs.centroid — Function

centroid(collection) -> eltype(collection)

Get the centroid of a collection of points. The preferred point representation is a subtype of StaticVector. This centroid is an unweighted average of positions.

ProtStructs.centered — Function

centered(collection) -> centered positions

Get a copy of the centered positions in collection. The preferred point representation is a subtype of StaticVector.

Equivalent to collection .- centroid(collection).

ProtStructs.RMSD — Function

RMSD(cA, cB) -> Real

Calculate root-mean-square deviation of points in cA from those in cB. Preferred point representation is a subtype of StaticVector.

Related constants

ProtStructs.AminoAcidNames — Constant

AminoAcidNames

A Dict of three-letter amino acid residue codes and their corresponding one-letter codes.

ProtStructs.DSSPConstants — Constant

Protstructs.DSSPConstants

Constants used for calculations in DSSP source code. (will I get in trouble for sharing this?). This Dict defines the following:

Minimum acceptable H–bond interaction distance: $0.5$ (angstroms?)

Minimum distance between alpha carbons: $9.0$ (angstroms?)

Maximum N–C distance in a peptide bond: $2.5$ (angstroms?)

HBond energy range: $[-9.9, -0.5]$ (kcal/mol)

Coupling constant: $-332 * 0.42 * 0.2$ (appropriate units)

Note: I may get in trouble for sharing this. I thought it would be fine since it's on the github hosted version.